backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asgiref>=3.8.0",
    "autopep8>=2.0.0",
    "django>=5.0.0,<6.0",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.0.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.9",
    "pycodestyle>=2.11.0",
    "pyjwt>=2.8.0",
    "sqlparse>=0.5.0",
]
